fix: minor fixes
TICKET: WIN-4296
